progress in copernicus DEM fetch

+++ TopoPyScale/fetch_dem.py
@@ -10,7 +10,8 @@ TODO:
 
 """
 
-import sys, os
+import sys, os, tarfile
+from pathlib import Path
 import requests
 from xml.dom.minidom import parse, parseString
 import pandas as pd
@@ -20,55 +21,68 @@ import warnings, os
 warnings.filterwarnings('ignore')
 
 class copernicus_dem():
-    """Class to download Publicly available Copernicus DEM products
+    """---> Class to download Publicly available Copernicus DEM products
     
-    Online Resources:
-        - [Sentinel Online blog post](https://sentinels.copernicus.eu/web/sentinel/-/copernicus-dem-new-direct-data-download-access)
-        - [Airbus data description](https://spacedata.copernicus.eu/documents/20123/122407/GEO1988-CopernicusDEM-SPE-002_ProductHandbook_I5.0+%281%29.pdf)
-        - [tutorial](https://spacedata.copernicus.eu/documents/20123/121286/Copernicus+DEM+Open+HTTPS+Access.pdf)
-
     Args:
         directory (str): directory where to store downloads
     """
-    
 
-    def __init__(self, directory):
-        self.directory = directory
+    def __init__(self, directory='.'):
+        self.directory = Path(directory)
         self.copernicus_public_url = 'https://prism-dem-open.copernicus.eu/pd-desk-open-access/publicDemURLs'
+
+        print("""
+Class to download Publicly available Copernicus DEM products. 
+
+Data are directly available via HTTPS requests at https://prism-dem-open.copernicus.eu/pd-desk-open-access/publicDemURLs
+
+Further online Resources:
+    - Sentinel Online blog post: https://sentinels.copernicus.eu/web/sentinel/-/copernicus-dem-new-direct-data-download-access
+    - Airbus data description: https://spacedata.copernicus.eu/documents/20123/122407/GEO1988-CopernicusDEM-SPE-002_ProductHandbook_I5.0+%281%29.pdf
+    - tutorial: https://spacedata.copernicus.eu/documents/20123/121286/Copernicus+DEM+Open+HTTPS+Access.pdf
+""")
         
 
     def fetch_copernicus_dem_product_list(self):
-        self.r_prod = requests.get(self.copernicus_public_url)
-        products_xml = parseString(r_prod.content)
+        self.r_products = requests.get(self.copernicus_public_url)
+        products_xml = parseString(self.r_products.content)
         products = products_xml.getElementsByTagName('datasetId')
 
         self.product_list = []
         print('---> List of product available:')
-        for d in products:
+        for i, d in enumerate(products):
             self.product_list.append(d.firstChild.data)
-            print(f'\t- {d.firstChild.data}')
-
+            print(f'\t [{i}]-> {d.firstChild.data}')
+        self._product_ind = int(input('Which product to download?'))
             
-    def request_tile_list(self, product, fname_tile='tiles.csv'):
+    def request_tile_list(self, product=None, fname_tile=None):
         """Function to 
 
         Args:
             product (_type_): _description_
             fname_tile (str, optional): _description_. Defaults to 'tiles.csv'.
-
-        Returns:
-            _type_: _description_
         """
-        headers = {'accept': 'csv',}
-        url = f"{self.copernicus_public_url}/{product.replace('/', '__')}"
-        r = requests.get(url, headers=headers)
-        
-        with open(fname_tile, 'w') as file:
-            file.write(r.content.decode("utf-8"))
+        if product is None:
+            product = self.product_list[self._product_ind]
+        if fname_tile is None:
+            fname_tile = 'tar_list_' + self.product_list[self._product_ind].replace('/', '__') + '.csv'
+            
+        # Add code to check first if fname_file does not exist already
+        if not os.path.isfile(self.directory / fname_tile):
 
-        print(f'---> All tile URL written to {fname_tile}')
+            print(f'---> Requesting list of tile for {product}')
 
-        df = pd.read_csv(fname_tile, header=None)
+            headers = {'accept': 'csv',}
+            url = f"{self.copernicus_public_url}/{product.replace('/', '__')}"
+            r = requests.get(url, headers=headers)
+            
+            with open(fname_tile, 'w') as file:
+                file.write(r.content.decode("utf-8"))
+
+            print(f'---> All tile URL written to {fname_tile}')
+
+        # open csv file containing list of tar file urls
+        df = pd.read_csv(self.directory / fname_tile, header=None)
         df.columns=['url']
 
         # extract latitude and longitude of tiles from filename
@@ -99,13 +113,56 @@ class copernicus_dem():
         import urllib.request as ur
         
         sub = self.df_tile_list.loc[(self.df_tile_list.lon>=extent[0]) & (self.df_tile_list.lon<=extent[1]) & (self.df_tile_list.lat>=extent[2]) & (self.df_tile_list.lat<=extent[3])]
-        
+
+        tar_list = []
         for i, row in sub.iterrows():
             print(f'---> Downloading {row.url}')
-            ur.urlretrieve(row.url, f"{self.directory}/{row.url.split('/')[-1]}")
-        
+            tar_file = self.directory / row.url.split('/')[-1]
+            ur.urlretrieve(row.url, tar_file)
+            tar_list.append(tar_file)
+
+        sub['tar_file'] = tar_list
         self.df_downloaded = sub
 
+    def _extract_tar(tar_obj, file_to_extract, target_dir='.'):
+        """Function extract a file from a .tar archive
+
+        Args:
+            tar_obj (tarfile obj): tarfile object
+            file_to_extract (_type_): _description_
+            target_dir (str, optional): _description_. Defaults to '.'.
+        """
+        
+        file = tar_obj.extractfile(file_to_extract)
+
+        p = Path(target_dir)
+        print(f'---> Extracting file {(p / file_to_extract).name} \tfrom {fname}')
+
+        with open((p / file_to_extract).name, 'wb') as outfile:
+            outfile.write(file.read())
+        file = None
+        
+
+    def extract_all_tar(self):
+        print('Not finished: does not find dt2 file as is')
+        
+        for i, row in self.df_downloaded.iterrows():
+            
+            tar_file = Path(row.url).name
+            tar = tarfile.open(tar_file, "r")
+            file = None
+            for tarinfo in tar.getmembers():
+                if tarinfo.name[-3:] == 'dt2':
+                    file = tarinfo.name
+            if file is not None:
+                self._extract_tar(tar, file_to_extract=file)
+            else:
+                raise ValueError('No .dt2 file in tar archive')
+            
+        print('\n---> READ DATA LICENSING IN TAR FILE <---')
+        print('--->        eula_F.pdf               <---\n')
+        
+
